requirements/local.txt
-r base.txt

codecov==2.0.15
django-coverage-plugin==1.5.0
django-debug-toolbar==1.9.1
django-webtest==1.9.2
factory-boy==2.10.0
faker==0.8.13
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-sugar==0.9.1
pytest==3.5.0
tox==3.0.0

